fixed photo not deleting on edit

+++ Pages/Admin/Products/Edit.cshtml.cs
@@ -38,7 +38,6 @@ namespace ComputerECommerce.Pages.Admin.Products
             ProductDto.Description = product.Description;
             ProductDto.Price = product.Price;
             ProductDto.Quantity = product.Quantity;
-            //ProductDto.Category = product.Category;
 
             Product = product;
         }
@@ -71,10 +70,15 @@ namespace ComputerECommerce.Pages.Admin.Products
                 {
                     ProductDto.ImageFile.CopyTo(fileStream);
                 }
-                string oldFilePath = Path.Combine(env.WebRootPath, product.Image);
-                if (System.IO.File.Exists(oldFilePath))
+                if(product.Image != "/images/products/laptop.jpg")
                 {
-                    System.IO.File.Delete(oldFilePath);
+                    string oldFilePath = env.WebRootPath + product.Image;
+                    Debug.WriteLine("\n\n\nTrying to delete file: " + oldFilePath);
+                    if(System.IO.File.Exists(oldFilePath))
+                    {
+                        Debug.WriteLine("Deleting file: " + oldFilePath);
+                        System.IO.File.Delete(oldFilePath);
+                    }
                 }
                 newFileName = "/images/products/" + newFileName;
             }
